Prepare next into dev

+++ src/Spinner/Core/WidthDeterminer.php
@@ -22,9 +22,12 @@ final class WidthDeterminer
         if (null === self::$determiner) {
             self::$determiner = self::createDeterminer();
         }
-        return (self::$determiner)($string);
+        return (int)(self::$determiner)($string);
     }
 
+    /**
+     * @codeCoverageIgnore
+     */
     private static function createDeterminer(): Closure
     {
         if (function_exists('\AlecRabbit\WCWidth\wcswidth')) {
